+++ crates/full-node/sov-sequencer/tests/integration/utils.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use proptest::bits::u64;
 use sov_chain_state::ChainState;
-use sov_mock_da::storable::service::StorableMockDaService;
+use sov_mock_da::storable::StorableMockDaService;
 use sov_mock_da::{BlockProducingConfig, MockAddress, MockDaService};
 use sov_mock_zkvm::crypto::private_key::Ed25519PrivateKey;
 use sov_mock_zkvm::MockZkvm;
